+++ server/internal/remote/clipboard/clipboard.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <libclipboard.h>
+#include <string.h>
+
+clipboard_c *getClipboard(void);
+
+void ClipboardSet(char *src);
+char *ClipboardGet();
