+++ server/internal/remote/xorg/xorg.c
@@ -1,6 +1,5 @@
 #include "xorg.h"
 
-static clipboard_c *CLIPBOARD = NULL;
 static Display *DISPLAY = NULL;
 static char *NAME = ":0.0";
 static int REGISTERED = 0;
@@ -73,13 +72,6 @@ Display *getXDisplay(void) {
   }
 
   return DISPLAY;
-}
-
-clipboard_c *getClipboard(void) {
-  if (CLIPBOARD == NULL) {
-    CLIPBOARD = clipboard_new(NULL);
-  }
-  return CLIPBOARD;
 }
 
 void XDisplayClose(void) {
@@ -206,16 +198,6 @@ void XKey(KeySym key, int down) {
 
   XTestFakeKeyEvent(display, code, down, CurrentTime);
   XSync(display, 0);
-}
-
-void XClipboardSet(char *src) {
-  clipboard_c *cb = getClipboard();
-  clipboard_set_text_ex(cb, src, strlen(src), 0);
-}
-
-char *XClipboardGet() {
-  clipboard_c *cb = getClipboard();
-  return clipboard_text_ex(cb, NULL, 0);
 }
 
 void XGetScreenConfigurations() {
